src/requirements.txt
fastapi>=0.95,<1
uvicorn[standard]>=0.22,<1

# Async DB
sqlalchemy[asyncio]>=2.0,<3
aiomysql>=0.1.0

# Pydantic v2 + settings helper
pydantic>=2.5,<3
pydantic-settings>=2.5,<3

# HTTP client for async calls to payment providers
httpx>=0.24,<1

# Migrations
alembic>=1.12,<2

# Optional helpers
python-dotenv>=1.0,<2

# Testing
pytest>=7.0,<8
pytest-asyncio>=0.22,<1